requirements.txt
asgiref==3.7.2
attrs==23.1.0
build==0.10.0
CacheControl==0.12.14
certifi==2023.5.7
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==3.1.0
cleo==2.0.1
crashtest==0.4.1
cryptography==41.0.1
defusedxml==0.7.1
distlib==0.3.6
Django==4.2.2
django-cors-headers==4.1.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
drf-extra-fields==3.5.0
drf-spectacular==0.26.3
dulwich==0.21.5
filelock==3.12.2
filetype==1.2.0
flake8==6.0.0
Flake8-pyproject==1.2.3
html5lib==1.1
identify==2.5.24
idna==3.4
importlib-metadata==6.7.0
inflection==0.5.1
installer==0.7.0
isort==5.12.0
jaraco.classes==3.2.3
jsonschema==4.17.3
keyring==23.13.1
lockfile==0.12.2
mccabe==0.7.0
more-itertools==9.1.0
msgpack==1.0.5
nodeenv==1.8.0
oauthlib==3.2.2
packaging==23.1
pexpect==4.8.0
Pillow==10.0.0
pkginfo==1.9.6
platformdirs==3.8.0
poetry==1.5.1
poetry-core==1.6.1
poetry-plugin-export==1.4.0
pre-commit==3.3.3
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyJWT==2.7.0
pyproject_hooks==1.0.0
pyrsistent==0.19.3
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
rapidfuzz==2.15.1
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
shellingham==1.5.0.post1
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
tomli==2.0.1
tomlkit==0.11.8
trove-classifiers==2023.5.24
typing_extensions==4.6.3
uritemplate==4.1.1
urllib3==1.26.16
virtualenv==20.23.1
webencodings==0.5.1
xattr==0.10.1
zipp==3.15.0
